Move identity tp to block_info/python
This commit is part of the process of rewriting
the identity tp in rust. It moves the python
version of identity tp from families/identity/
to families/identity/python.
This commit also updates all references to
files that are part of the identity tp
to point to the new path. A future commit will add
the rust version of identity at families/identity/rust.

Signed-off-by: Eloá Franca Verona <eloafran@bitwise.io>